- fix: init v2
