+++ ldapper_test.go
@@ -171,16 +171,16 @@ func TestTestBaseDn(t *testing.T) {
 	}
 }
 
-func TestGetUserInfo(t *testing.T) {
+func TestReadUserInfo(t *testing.T) {
 	type ConnParams struct {
-		host       string
-		port       interface{}
-		user       string
-		password   string
-		useTLS     bool
-		openLDAP   bool
-		baseDN     string
-		userToFind string
+		host     string
+		port     interface{}
+		user     string
+		domUser  string
+		password string
+		useTLS   bool
+		openLDAP bool
+		baseDN   string
 	}
 
 	tests := []struct {
@@ -191,26 +191,22 @@ func TestGetUserInfo(t *testing.T) {
 		{
 			name: "invalid notls",
 			params: ConnParams{
-				host:       "test.ru",
-				port:       389,
-				user:       "test",
-				password:   "test",
-				useTLS:     false,
-				baseDN:     "OU=test,DC=test,DC=ru",
-				userToFind: "test",
+				host:     "test.ru",
+				port:     389,
+				user:     "test",
+				password: "test",
+				useTLS:   false,
 			},
 			mustFail: true,
 		},
 		{
 			name: "invalid tls",
 			params: ConnParams{
-				host:       "test.ru",
-				port:       636,
-				user:       "test",
-				password:   "test",
-				useTLS:     true,
-				baseDN:     "OU=test,DC=test,DC=ru",
-				userToFind: "test",
+				host:     "test.ru",
+				port:     636,
+				user:     "test",
+				password: "test",
+				useTLS:   true,
 			},
 			mustFail: true,
 		},
@@ -218,24 +214,231 @@ func TestGetUserInfo(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			conn, err := NewLdapConn(tt.params.user, tt.params.password,
+			_, err := ReadUserInfo(tt.params.user, tt.params.domUser, tt.params.password,
 				tt.params.host, tt.params.port,
-				tt.params.useTLS, LdapConnOptions{openLDAP: tt.params.openLDAP})
+				tt.params.baseDN, tt.params.useTLS, tt.params.openLDAP)
 			if tt.mustFail {
 				require.Error(t, err)
 			} else {
 				require.NoError(t, err)
-				defer func() { conn.Close() }()
-
-				res, err := conn.GetUserInfo(tt.params.userToFind, tt.params.baseDN)
-				if tt.mustFail {
-					require.Error(t, err)
-				} else {
-					require.NoError(t, err)
-					t.Log(res)
-				}
 			}
+		})
+	}
+}
 
+func TestReadRootGroups(t *testing.T) {
+	type ConnParams struct {
+		host     string
+		port     interface{}
+		user     string
+		password string
+		useTLS   bool
+		openLDAP bool
+		baseDN   string
+	}
+
+	tests := []struct {
+		name     string
+		params   ConnParams
+		mustFail bool
+	}{
+		{
+			name: "invalid notls",
+			params: ConnParams{
+				host:     "test.ru",
+				port:     389,
+				user:     "test",
+				password: "test",
+				useTLS:   false,
+			},
+			mustFail: true,
+		},
+		{
+			name: "invalid tls",
+			params: ConnParams{
+				host:     "test.ru",
+				port:     636,
+				user:     "test",
+				password: "test",
+				useTLS:   true,
+			},
+			mustFail: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			_, err := ReadRootGroups(tt.params.user, tt.params.password,
+				tt.params.host, tt.params.port,
+				tt.params.baseDN, tt.params.useTLS, tt.params.openLDAP)
+			if tt.mustFail {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+			}
+		})
+	}
+}
+
+func TestReadSubGroups(t *testing.T) {
+	type ConnParams struct {
+		host     string
+		port     interface{}
+		user     string
+		password string
+		useTLS   bool
+		openLDAP bool
+		group    string
+		level    int
+	}
+
+	tests := []struct {
+		name     string
+		params   ConnParams
+		mustFail bool
+	}{
+		{
+			name: "invalid notls",
+			params: ConnParams{
+				host:     "test.ru",
+				port:     389,
+				user:     "test",
+				password: "test",
+				useTLS:   false,
+			},
+			mustFail: true,
+		},
+		{
+			name: "invalid tls",
+			params: ConnParams{
+				host:     "test.ru",
+				port:     636,
+				user:     "test",
+				password: "test",
+				useTLS:   true,
+			},
+			mustFail: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			_, err := ReadSubGroups(tt.params.user, tt.params.password, tt.params.group,
+				tt.params.level, tt.params.host, tt.params.port,
+				tt.params.useTLS, tt.params.openLDAP)
+			if tt.mustFail {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+			}
+		})
+	}
+}
+
+func TestReadGroupUsers(t *testing.T) {
+	type ConnParams struct {
+		host     string
+		port     interface{}
+		user     string
+		password string
+		useTLS   bool
+		openLDAP bool
+		group    string
+	}
+
+	tests := []struct {
+		name     string
+		params   ConnParams
+		mustFail bool
+	}{
+		{
+			name: "invalid notls",
+			params: ConnParams{
+				host:     "test.ru",
+				port:     389,
+				user:     "test",
+				password: "test",
+				useTLS:   false,
+			},
+			mustFail: true,
+		},
+		{
+			name: "invalid tls",
+			params: ConnParams{
+				host:     "test.ru",
+				port:     636,
+				user:     "test",
+				password: "test",
+				useTLS:   true,
+			},
+			mustFail: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			_, err := ReadGroupUsers(tt.params.user, tt.params.password, tt.params.group,
+				tt.params.host, tt.params.port,
+				tt.params.useTLS, tt.params.openLDAP)
+			if tt.mustFail {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+			}
+		})
+	}
+}
+
+func TestReadAdStruct(t *testing.T) {
+	type ConnParams struct {
+		host     string
+		port     interface{}
+		user     string
+		password string
+		useTLS   bool
+		openLDAP bool
+		baseDN   string
+	}
+
+	tests := []struct {
+		name     string
+		params   ConnParams
+		mustFail bool
+	}{
+		{
+			name: "invalid notls",
+			params: ConnParams{
+				host:     "test.ru",
+				port:     389,
+				user:     "test",
+				password: "test",
+				useTLS:   false,
+			},
+			mustFail: true,
+		},
+		{
+			name: "invalid tls",
+			params: ConnParams{
+				host:     "test.ru",
+				port:     636,
+				user:     "test",
+				password: "test",
+				useTLS:   true,
+			},
+			mustFail: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			_, err := ReadAdStruct(tt.params.user, tt.params.password,
+				tt.params.host, tt.params.port,
+				tt.params.baseDN, tt.params.useTLS, tt.params.openLDAP)
+			if tt.mustFail {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+			}
 		})
 	}
 }
