--- pack_init.go
@@ -1,9 +1,0 @@
-// Copyright 2020-2024 NGR Softlab
-//
-// ldap_common - pack with ldap functions (supports ldap and openLdap now (tls & noTls)).
-//
-// creator:	everyone
-// using: ldapper "github.com/NGRsoftlab/ngr-ldapper"
-package ldapper
-
-/////////////////////////////////////////////
