+++ v2/ldapper_test.go
@@ -12,16 +12,13 @@ import (
 	To check real work - you need to put real params into test case (ip, port, user, pass, etc.)
 */
 
-func TestReadUserInfo(t *testing.T) {
+func TestNewLdapConn(t *testing.T) {
 	type ConnParams struct {
 		host     string
 		port     interface{}
 		user     string
-		domUser  string
 		password string
 		useTLS   bool
-		openLDAP bool
-		baseDN   string
 	}
 
 	tests := []struct {
@@ -55,27 +52,26 @@ func TestReadUserInfo(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			_, err := ReadUserInfo(tt.params.user, tt.params.domUser, tt.params.password,
+			conn, err := NewLdapConn(tt.params.user, tt.params.password,
 				tt.params.host, tt.params.port,
-				tt.params.baseDN, tt.params.useTLS, tt.params.openLDAP)
+				tt.params.useTLS)
 			if tt.mustFail {
 				require.Error(t, err)
 			} else {
+				defer func() { conn.Close() }()
 				require.NoError(t, err)
 			}
 		})
 	}
 }
 
-func TestReadRootGroups(t *testing.T) {
+func TestTryAccess(t *testing.T) {
 	type ConnParams struct {
 		host     string
 		port     interface{}
 		user     string
 		password string
 		useTLS   bool
-		openLDAP bool
-		baseDN   string
 	}
 
 	tests := []struct {
@@ -109,9 +105,9 @@ func TestReadRootGroups(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			_, err := ReadRootGroups(tt.params.user, tt.params.password,
+			err := TryAccess(tt.params.user, tt.params.password,
 				tt.params.host, tt.params.port,
-				tt.params.baseDN, tt.params.useTLS, tt.params.openLDAP)
+				tt.params.useTLS)
 			if tt.mustFail {
 				require.Error(t, err)
 			} else {
@@ -121,116 +117,7 @@ func TestReadRootGroups(t *testing.T) {
 	}
 }
 
-func TestReadSubGroups(t *testing.T) {
-	type ConnParams struct {
-		host     string
-		port     interface{}
-		user     string
-		password string
-		useTLS   bool
-		openLDAP bool
-		group    string
-		level    int
-	}
-
-	tests := []struct {
-		name     string
-		params   ConnParams
-		mustFail bool
-	}{
-		{
-			name: "invalid notls",
-			params: ConnParams{
-				host:     "test.ru",
-				port:     389,
-				user:     "test",
-				password: "test",
-				useTLS:   false,
-			},
-			mustFail: true,
-		},
-		{
-			name: "invalid tls",
-			params: ConnParams{
-				host:     "test.ru",
-				port:     636,
-				user:     "test",
-				password: "test",
-				useTLS:   true,
-			},
-			mustFail: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := ReadSubGroups(tt.params.user, tt.params.password, tt.params.group,
-				tt.params.level, tt.params.host, tt.params.port,
-				tt.params.useTLS, tt.params.openLDAP)
-			if tt.mustFail {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-			}
-		})
-	}
-}
-
-func TestReadGroupUsers(t *testing.T) {
-	type ConnParams struct {
-		host     string
-		port     interface{}
-		user     string
-		password string
-		useTLS   bool
-		openLDAP bool
-		group    string
-	}
-
-	tests := []struct {
-		name     string
-		params   ConnParams
-		mustFail bool
-	}{
-		{
-			name: "invalid notls",
-			params: ConnParams{
-				host:     "test.ru",
-				port:     389,
-				user:     "test",
-				password: "test",
-				useTLS:   false,
-			},
-			mustFail: true,
-		},
-		{
-			name: "invalid tls",
-			params: ConnParams{
-				host:     "test.ru",
-				port:     636,
-				user:     "test",
-				password: "test",
-				useTLS:   true,
-			},
-			mustFail: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := ReadGroupUsers(tt.params.user, tt.params.password, tt.params.group,
-				tt.params.host, tt.params.port,
-				tt.params.useTLS, tt.params.openLDAP)
-			if tt.mustFail {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-			}
-		})
-	}
-}
-
-func TestReadAdStruct(t *testing.T) {
+func TestTestBaseDn(t *testing.T) {
 	type ConnParams struct {
 		host     string
 		port     interface{}
@@ -272,7 +159,7 @@ func TestReadAdStruct(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			_, err := ReadAdStruct(tt.params.user, tt.params.password,
+			err := TestBaseDn(tt.params.user, tt.params.password,
 				tt.params.host, tt.params.port,
 				tt.params.baseDN, tt.params.useTLS, tt.params.openLDAP)
 			if tt.mustFail {
@@ -280,6 +167,75 @@ func TestReadAdStruct(t *testing.T) {
 			} else {
 				require.NoError(t, err)
 			}
+		})
+	}
+}
+
+func TestGetUserInfo(t *testing.T) {
+	type ConnParams struct {
+		host       string
+		port       interface{}
+		user       string
+		password   string
+		useTLS     bool
+		openLDAP   bool
+		baseDN     string
+		userToFind string
+	}
+
+	tests := []struct {
+		name     string
+		params   ConnParams
+		mustFail bool
+	}{
+		{
+			name: "invalid notls",
+			params: ConnParams{
+				host:       "test.ru",
+				port:       389,
+				user:       "test",
+				password:   "test",
+				useTLS:     false,
+				baseDN:     "OU=test,DC=test,DC=ru",
+				userToFind: "test",
+			},
+			mustFail: true,
+		},
+		{
+			name: "invalid tls",
+			params: ConnParams{
+				host:       "test.ru",
+				port:       636,
+				user:       "test",
+				password:   "test",
+				useTLS:     true,
+				baseDN:     "OU=test,DC=test,DC=ru",
+				userToFind: "test",
+			},
+			mustFail: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			conn, err := NewLdapConn(tt.params.user, tt.params.password,
+				tt.params.host, tt.params.port,
+				tt.params.useTLS, LdapConnOptions{openLDAP: tt.params.openLDAP})
+			if tt.mustFail {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+				defer func() { conn.Close() }()
+
+				res, err := conn.GetUserInfo(tt.params.userToFind, tt.params.baseDN)
+				if tt.mustFail {
+					require.Error(t, err)
+				} else {
+					require.NoError(t, err)
+					t.Log(res)
+				}
+			}
+
 		})
 	}
 }
